Split image-tracking-target into its own file and add documentation
